Refactor to use Enum as output data object.
