+++ core/src/main/java/com/aarves/bluepages/usecase/data/AccountDataImpl.java
@@ -2,7 +2,7 @@ package com.aarves.bluepages.usecase.data;
 
 import com.aarves.bluepages.entities.User;
 import com.aarves.bluepages.usecase.interactors.AccountData;
-import com.aarves.bluepages.usecase.interactors.PermissionsFailureException;
+import com.aarves.bluepages.usecase.exceptions.PermissionsFailureException;
 import com.aarves.bluepages.usecase.interactors.ReviewRepository;
 
 public class AccountDataImpl implements AccountData {
@@ -26,7 +26,11 @@ public class AccountDataImpl implements AccountData {
 
     @Override
     public User getUserAccount(String username, String passwordHash) throws PermissionsFailureException {
-        if(this.isExistingAccount(username) && this.isPasswordMatch(username, passwordHash)) {
+        if(!this.isExistingAccount(username)) {
+            return null;
+        }
+
+        if(this.isPasswordMatch(username, passwordHash)) {
             User user = new User(username, passwordHash);
             user.setReviews(this.reviewRepository.getReviewsByUser(user));
