+++ core/src/main/java/com/aarves/bluepages/usecase/exceptions/NotLoggedInException.java
@@ -1,4 +1,4 @@
-package com.aarves.bluepages.usecase.interactors;
+package com.aarves.bluepages.usecase.exceptions;
 
 public class NotLoggedInException extends Exception {
 }
