+++ core/src/main/java/com/aarves/bluepages/usecase/exceptions/PermissionsFailureException.java
@@ -1,4 +1,4 @@
-package com.aarves.bluepages.usecase.interactors;
+package com.aarves.bluepages.usecase.exceptions;
 
 public class PermissionsFailureException extends Exception {
 }
