+++ core/src/main/java/com/aarves/bluepages/usecase/interactors/AccountPresenter.java
@@ -1,6 +1,6 @@
 package com.aarves.bluepages.usecase.interactors;
 
 public interface AccountPresenter {
-    void loginResult(boolean result, String username);
+    void loginResult(LoginResult result, String username);
     void registerResult(boolean result);
 }
