+++ core/src/main/java/com/aarves/bluepages/usecase/interactors/LoginResult.java
@@ -1,0 +1,7 @@
+package com.aarves.bluepages.usecase.interactors;
+
+public enum LoginResult {
+    SUCCESS,
+    FAILURE,
+    ACCOUNT_NOT_FOUND
+}
